feat: refactored package
